Initial work for the game pager
Sepearted the parts into fragments
TODO complete layout for the game controls
Prevent users from swiping away the app? If that isn't
done then I will need to implement a feature where the
phone reopens the app.  Or I can add a simple clock to the
app, so they always stay in the app