A lot of work to the project
Finish activity is now complete, adjusted the mute mode for phone
Will set it so if the app detects an android wear it will turn on
wear mode, which will mute the phone, and display a message to the user.
The user can then disable or keep it.
Wear mode mutes the phone and disables the always on screen.