Add global game information state
If the user closes the watch app, and the minute is up.  The app
will be relaunched, and shown the shot notification.

The showing of the SHOT notification and animation is not yet working
but all the information is there.

This will also help with the ongoing Notification I will be adding
later